# Performance of MovingMinMax0 is now slightly better.