I optimized MovingMinMax0, on my test machine, it is now consistently faster than MovingMinMax by a tiny amount (~3%) suggesting that we may have reached something close to the best you can do in pure Go (maybe?).

Note how computing the max, and the min separately costs a lot more than computing the min and max together!

     $ go test -bench=.
     PASS
     BenchmarkReference	   50000	     50037 ns/op
     BenchmarkMovingMinMax0	  100000	     25181 ns/op
     BenchmarkMovingMinMax	  100000	     26129 ns/op
     BenchmarkMovingMin	  100000	     18446 ns/op
     BenchmarkMovingMax	  100000	     17971 ns/op
     ok  	_/home/dlemire/CVS/github/movingminmax	13.129s